spread pattern fix
tweak_data.weapon.spread.dispersion is defaulted to 1.0 which distributes shots evenly by radius. However by area shots will be weighted towards the middle of your spread, if you want spread to distribute evenly by area change this value to 0.5. Just note this will have the opposite effect and shots will be weighted to the outside of your spread radius.

Personally I think 1.0 is the better value as the radius of spread is what makes your shot hit or miss but it is visually weighted to the center.