Attempted work on S&W .500
crashes ahoy! Feel free to try fixing it
all the assets should be there now with all of the under the id jeb instead of having to switch back and forth.
Commented out the lua and xml code to prevent crashes.
The weaponfactorytweadata.lua for Jeb should be complete now.