XM25 smart scope fix + airburst standalone
Fixed incendiary grenades getting the airburst function. Now you can use the incendiary grenade with the XM25 scope, but it will not have the airburst function.
Also made a separate standalone attachment for those who might want the airburst function but without the scope due to issues (including part of the player hand covering up the scope view). You need to buy the attachment with the coins due to it being not available by default for some reason.