migrates most of my heist edits to the new replacement sys.
Actual Changes:
- Moves my garbage heist edits to the new replacement system, so that I can berated even more for being an idiot.
- Also adds in a small Birth of Sky edit, which makes actual murkydozers spawn in the plane and disables sewer spawns. (might've fucked the scripting for the sewer part, pls test, kthx!!!)

NOTES/TIDBITS:
I also wanted to make the plane spawns on Birth of Sky difficulty specific, with the first part of the plane having our green murkies and then the upper part of the plane having elite murkies. I couldn't do that since I'm not that good with beardlib editor atm. if anyone else wants to take a crack at it, the instance paths are:
assets/levels/instances/unique/pbr/pbr_plane_cargo (for the lower plane part, has most of the murky spawns excluding the dozers.)
assets/levels/instances/unique/pbr/pbr_plane_crew (for the rest of the spawns and the upper level of the plane)