Improved vanilla bot enemy marking a bit + turret marking
Regarding the improvements, now it checks if the unit causes a reaction (and a combat one) before anything else, then checks if it can use contours. Then, proceeds as usual, but also checks if the unit isn't already marked (using red/pink contours, so that Medic healing contours won't invalidate this) before actually deciding if it's a valid unit to mark based on closest distance.