Testing sync time of day
When adding new environments to time of day, make sure to add the menu option for it in main.xml (around line 349 at the time of writing this), and then add the setting name to the list of sync-able environments, in Core.lua (around line 132), in table environment_settings.
Added a Hook that is automatically called whenever another client is fully synced to oneself.
On peer sync complete, the host will send their time of day settings to all other clients. The clients will receive this data and save it to (over) their own copies of Environment_Settings_Table, though it won't be called yet. Probably needs to call `TODCallBeardLibSequenceFuncs` hook or something.
Also I added some debug logs in Core and entry, those should be removed once this is verified working.