Mind Blown-Up
Mind Blown now applies to single fire smgs and single fire rifles in addition to snipers. They don't benefit much from the full 1.0 damage mult on kill when aced so it's unlikely to break balance, and it adds some consistency to the tree.
Mind Blown now chains to a single enemy, and then up to 4 additional enemies the further away the player is.
Killing headshots chain an additional time if aced.