Improved how poison DOTs are applied + syncing
So basically now DOT is synced instead of just the damage itself being synced between players. This prevents poison stacking on enemies the more peers there are in a game. Worry not, damage is still done locally by the attacking peer, making the mechanic almost identical and consistent with fire doting.