wave count fix
There is a problem with the huds wave counter where it doesn't seem to show the correct wave number while joining a lobby as a client.

I made some tweaks so it will show the correct wave when dropping in and also update correctly when fighting off assaults both in safehouse raid and holdout.

The reason I had it be hidden when the assault end and shown when the next assault starts was because it kept showing the current wave number as 0 until the next wave started. So it was a work around for that.