Sanity check + use unused synced crouch/stand pos
I also added a debug toggle to see when and where is the position being updated (don't expect it to be accurate with the husk's head model as it's not accurate to a player's camera anyway). Also do note that the center of the sphere is the actual position, the sphere is not what's being used for detection.