Weapon Swap Speed Stuff
Weapon Swap Speed now scales off of the concealment stat, and is displayed along with other weapon stats in the black market.
The displayed stat is the combined equip and unequip delays. They're usually pretty close on a lot of weapons.

A concealment of 20 means that the swap speed timer in unaffected.
Going above 20 gives you up to a 1.25 multiplier on swap speed.
Going below 20 gives you a multiplier down to 0.5 swap speed.
This multiplier scales linearly with concealment.
This means that concealment is a relevant stat to weight when building your guns, even outside of dodge builds.
You might shave a second off of that minigun equip and unequip if you get concealment mods!
Values chosen such that it has minimal effects on most existing weapon stats.

Innate bonus to swap speed given to pistols was moved to tweakdata, and is no longer tied to the skill system.