Fix for Turret shield damage syncing + polishing
The issue was a typo on my part, I made a certain variable a local twice, causing the first instance to be nil since it was in another code block.

The other addition is that if the shield is destroyed locally, all three damage types can force the shield to be destroyed for other peers. This ensures consistency with how the Turret getting destroyed works, which is similar to how enemies dying is handled.

So basically, there should now be 0 cases of a Turret not getting it's shield destroyed for some peers but not for others, even if the shield health values are altered locally.