Client-sided detection is now fully functional
There were some issues with casing mode. The system is so hard to get a full grasp on that it threw a curve on what I knew about it (to be more specific, being identified only means you're detected outside of casing mode, while in said mode, it requires being identified and then separately build up suspicion, I hate it).