Player husk position updating improvements
m_com (center of mass) and m_stand_pos are now consistent with how local players define them (and also makes the second get updated for player husks in the first place, as it never does in vanilla).

Retrieving m_pos() and m_head_pos() + any of the other mutables from this file is now based on the latest synced positions, rather than using positions based on the model. This was already done for detection before this change, but now it should apply globally. Visual movement should also not be affected by these changes.