Fixed a specific detection issue when masking up
Since both being masked up or not use notice_progress, but only the latter uses uncover_progress for suspicion and to detect the player, notice_progress needs to be reset if it hasn't reached 100% (which makes you identified), else the detection will jump to whatever point it was. When I made the meter not reset after masking up in general, I didn't think of this specific case (you need some timing since being noticed in casing mode is super fast).

Plus a sanity check that's normally not needed, but who knows later on.