CS Mission Rewards
Increased the rewards the missions give across the board, most heists will see an increase of 2-5 points, and aims to balance around difficulty AND length. The increases are also to help get players to higher levels of crime spree faster to avoid boredom at the lower levels.

Transport heists (except train) are unchanged and remain at 5 due to how incredibly short they are.