Fixed navigation around the office-roof and roof-office
Cops are unable to open the door so the only path for a cop to get from the roof to the office if the door is shut is through a badly bottlenecked singular navlink and the only way for a cop to go from the office to the roof is to take a detour to a staircase. The team AI are even worse and would always take a detour to one of the staircases unless the door is open.
Team AI will now always take a direct path between the office and roof instead of always taking a detour and cops are now able to breach the office door.