Fix an issue with path simplification
Paths wouldn't be simplified on slopes when they should be. When checking whether a navpoint can be removed they subtract the z differential of the current-next navpoint from the z differential between the current-previous (current.z - prev.z - (current.z - next.z)) But as they subtract instead of add this merely gets the z difference between the previous-next, the current has no influence on the end result at all.

If this is switched to adding instead of subtracting (like is almost certainly intended, since there's no reason to set it up this way instead of just doing prev.z - next.z) the current navpoint's z is actually relevant for checking whether a shortcut is possible. Even if the previous and next both have a substantial z difference between the middle, as long as the difference is below 0.6m it's still possible to shortcut (for example: (5m - 0m + (5m - 10m) = 0m). This also fixes an issue where if both the previous and next have identical z values but the middle has a substantially different z value it would incorrectly attempt to remove the current navpoint (vanilla example: (5m - 10m - (5m - 10m) = 0m)

You can visualise this in 2D if you imagine three points with z values of 100, 200, 300. A line between the first-second-third will be straight, making the middle point essentially useless. The line doesn't have to be perfectly straight as there is some tolerance, as well as an additional check that there's valid navigation between the previous and next because the middle navpoint can still be off to either side of the previous/next but you get the idea.