Restored pre-U173 G3 reload anims
Replacing that awful and impossible (sideways flick on a rack-in/out type magwell, and that's not even accounting for the absolutely MASSIVE hands you'd need to have to use the button release without adjusting your grip, lmao) G3 reload with the pre-U173 one.

Sped up the animation to match the 2 second empty reload the new one has, which is what should have been done in the first place if the duration was the issue, Overkill >:^(

While it does look cracked out due to the whopping 90% speed boost, so does the new one.