Fix some bullet impact issues
For example this fixes an issue where if a bullet hits a dozer's armour and destroys it it'd play a blood impact. Also re-enabled bullet impacts for shots that did no damage since whoever wrote the original code was completely ignorant to the fact you still need a bullet impact if it was blocked by armour, like FBI heavies.