The great enemy animation xml unbugening
This will not be enabled yet as SuperBLT hasn't been updated to support this!

Properly enables reloads while moving by removing the no reloads flag, cops could already reload while moving here but this is how it should've been handled.

Makes medic heals an upper_body animation to fix several issues with it, mainly in the case of vanilla where it's used as if it is upper_body. Medics are now capable of healing while moving if desired.

Concussion hurt animations will now set the required flags to actually work properly, so hurt actions will end at the same time the animation does instead of only a static amount of time after it starts, as well as allowing for a new concussion animation to be started if the current one is nearly finished (70% done). The unit will also ragdoll if they die during the animation if needed.

Simplified the shield knockback animation state a ton, this doesn't change anything for Resmod but would enable an otherwise unused animation in vanilla.

ECM feedback animations now require the unit to fully end the animation before ending the hurt action for a smoother transition back to idle, and will exit to a standing pose as it's more suitable after ending the animation. This also fixes the dozer crouching due to ECM feedback.

Poison hurt animations will now loop to allow for units to be affected by it for longer durations, fixes to ECM feedback are also applicable here.